+++ src/FlaUILibrary/flaui/util/converter.py
@@ -1,3 +1,4 @@
+import re
 from typing import Any
 from FlaUILibrary.flaui.exception import FlaUiError
 
@@ -57,3 +58,13 @@ class Converter:
             return False
 
         return bool(value)
+
+    @staticmethod
+    def get_combobox_xpath_from_combobox_selectionitem_xpath(xpath: str) -> str:
+        is_combobox_selectionitem = True if "ComboBox" in xpath and "ComboBox" not in xpath.split("/")[-1] else False
+        if is_combobox_selectionitem:
+            matches = re.findall(r"/ComboBox.*?/", xpath)
+            s = matches[0] if matches is not [] else ""
+            result = f"{xpath.split(s)[0]}{s}"[:-1] if s else ""
+            return result
+        return ""