+++ src/FlaUILibrary/keywords/property.py
@@ -1,7 +1,9 @@
 from robotlibcore import keyword
 from FlaUILibrary.flaui.module import Property
 from FlaUILibrary.flaui.exception import FlaUiError
+from FlaUILibrary.flaui.enum import InterfaceType
 from FlaUILibrary.flaui.util.automationinterfacecontainer import AutomationInterfaceContainer
+from FlaUILibrary.flaui.util.converter import Converter
 
 
 class PropertyKeywords:  # pylint: disable=too-many-public-methods
@@ -45,8 +47,10 @@ class PropertyKeywords:  # pylint: disable=too-many-public-methods
         | IS_VALUE_PATTERN_SUPPORTED | Bool | True or False |
         | VALUE | String | The Value Property of Element |
         | IS_EXPAND_COLLAPSE_PATTERN_SUPPORTED | Bool | True or False |
-        | EXPAND_COLLAPSE_STATE | String | Collapsed or Expanded |        
-
+        | EXPAND_COLLAPSE_STATE | String | Collapsed or Expanded |   
+        | IS_SELECTION_ITEM_PATTERN_SUPPORTED | Bool | True or False |    
+        | IS_SELECTED | Bool | True or False |
+        
         Possible FlaUI-Errors:
         | Element could not be found by xpath        |
         | Pattern is not supported by given element  |
@@ -64,24 +68,38 @@ class PropertyKeywords:  # pylint: disable=too-many-public-methods
 
         """
         # pylint: enable=line-too-long
-
-        module = self._container.get_module()
-        element = module.get_element(identifier, msg=msg)
+        
         action_value = ""
-
         try:
             action_value = Property.Action[action.upper()]
-        except KeyError:
+        except KeyError or action_value in [Property.Action.MAXIMIZE_WINDOW, Property.Action.MINIMIZE_WINDOW, Property.Action.NORMALIZE_WINDOW]:
             FlaUiError.raise_fla_ui_error(FlaUiError.InvalidPropertyArgument)
+        
+        module = self._container.get_module()
+        
+        if action_value is Property.Action.IS_SELECTED:
+            # need expand parent ComboBox before to get ComboBox SelectionItem element
+            combobox_xpath = Converter.get_combobox_xpath_from_combobox_selectionitem_xpath(identifier)
+            if combobox_xpath:
+                combobox_element = module.get_element(combobox_xpath, InterfaceType.COMBOBOX, msg)
+                module.action(Property.Action.STAGE_FOR_COMBOBOX_SELECTIONITEM,
+                            Property.create_value_container(element=combobox_element, uia=module.identifier()),
+                            msg)
 
-        if action_value is (Property.Action.MAXIMIZE_WINDOW
-                            or Property.Action.MINIMIZE_WINDOW
-                            or action_value == Property.Action.NORMALIZE_WINDOW):
-            FlaUiError.raise_fla_ui_error(FlaUiError.InvalidPropertyArgument)
-
-        return module.action(action_value,
-                             Property.create_value_container(element=element, uia=module.identifier()),
-                             msg)
+        element = module.get_element(identifier, msg=msg)
+        property_value = module.action(action_value,
+                            Property.create_value_container(element=element, uia=module.identifier()), 
+                            msg)
+        
+        if action_value is Property.Action.IS_SELECTED:
+            combobox_xpath = Converter.get_combobox_xpath_from_combobox_selectionitem_xpath(identifier)
+            if combobox_xpath:
+                combobox_element = module.get_element(combobox_xpath, InterfaceType.COMBOBOX, msg)
+                module.action(Property.Action.STAGE_FOR_COMBOBOX_SELECTIONITEM,
+                            Property.create_value_container(element=combobox_element, uia=module.identifier()),
+                            msg)
+        
+        return property_value
 
     @keyword
     def get_background_color(self, identifier, msg=None):